refactor(discord): split actions.ts into action modules
actions.ts was 730 lines — one createDiscordActionHandler factory
wrapping a big switch over 30+ Discord API actions. Split into an
actions/ folder mirroring the telegram actions/ layout:
  - types     — DiscordActionContext + handler-map types
  - shared    — tryAction + resolveChannel + buildButtonRows
  - messaging — send/reply/edit/delete/pin/forward/react/buttons/schedule
  - media     — attachments, stickers, polls, locations, contacts, dice
  - chat-info — chat/member info, titles, pinned, history/search, Telegram
                no-ops

index.ts merges the maps and rebuilds createDiscordActionHandler with the
identical signature, the up-front channel resolution + needsChannel guard,
and per-instance scheduledMessages state. Importer updated.
